+++ bookstore-backend/src/main/java/com/mybookstore/bookstore/utils/orderutils/OrderWithItems.java
@@ -1,19 +1,18 @@
 package com.mybookstore.bookstore.utils.orderutils;
 
 import lombok.Data;
-import net.sf.json.JSONObject;
 
 import java.util.ArrayList;
 import java.util.List;
 
 @Data
 public class OrderWithItems {
-    private int id;
-    private int userId;
-    private Double totalPrice;
+    private Integer id;
+    private Integer userId;
+    private Integer totalPrice;
     private List<Item> items;
 
-    public OrderWithItems(int _id, int _userId, Double _totalPrice) {
+    public OrderWithItems(Integer _id, Integer _userId, Integer _totalPrice) {
         id = _id;
         userId = _userId;
         totalPrice = _totalPrice;
